refactor(daemon): split dispatch lifecycle modules
Share one enqueue preparation path between manual dispatch and auto-hop while preserving scheduler/store lock order and post-commit effects.
